debian-base: add local-apt-repository package
Add a local apt repository to the Debian base by using the package
local-apt-repository. Technically it is added on an early stage but
will not be active at this stage unless the nativeaptinstall bbclass
is patched to perform an update once it has been installed. So, for
now it just prepares this local repository without actually using it.

We need this local repository to place custom packages to protect
what has been installed by Yocto at a later stage.

But it could also be used for other use cases.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>